+++ src/relayer/Relayer.t.sol
@@ -95,7 +95,10 @@ contract RelayerTest is DSTest {
         assertTrue(oracleData.exists);
         assertEq(oracleData.lastUpdateValue, 0);
         assertEq(oracleData.tokenId, mockTokenId1);
-        assertEq(oracleData.minimumThresholdValue, mockTokenId1MinThreshold);
+        assertEq(
+            oracleData.minimumPercentageDeltaValue,
+            mockTokenId1MinThreshold
+        );
     }
 
     function test_CheckExistenceOfOracle() public {
@@ -342,28 +345,142 @@ contract RelayerTest is DSTest {
     function test_Execute_DoesNotUpdatesRatesInCollybusWhenDeltaIsBelowThreshold()
         public
     {
-        // Execute must call update on all oracles before pushing the values to Collybus
-        relayer.execute();
+        // Threshold percentage
+        uint256 thresholdPercentage = 50_00; // 50%
 
-        // Make the second value returned by the oracle to be just lower than the minimum threshold
-        int256 oracleNewValue = oracle1InitialValue +
-            int256(mockTokenId1MinThreshold) -
-            1;
+        // TokenId
+        bytes32 localTokenId = bytes32("percentage_threshold_token_test");
 
-        oracle1.givenQueryReturnResponse(
+        // Create a local relayer
+        Relayer localRelayer = new Relayer(address(collybus), relayerType);
+
+        // Create a local oracle
+        MockProvider localOracle = new MockProvider();
+
+        // Set the value returned by the Oracle.
+        int256 initialValue = 100;
+        localOracle.givenQueryReturnResponse(
             abi.encodePacked(IOracle.value.selector),
             MockProvider.ReturnData({
                 success: true,
-                data: abi.encode(oracleNewValue, true)
+                data: abi.encode(initialValue, true)
             }),
             false
         );
 
-        relayer.execute();
+        // Add oracle with a thresold percentage
+        localRelayer.oracleAdd(
+            address(localOracle),
+            localTokenId,
+            thresholdPercentage
+        );
 
-        // The rate will NOT be updated because the delta is smaller than the threshold
-        assertTrue(
-            collybus.valueForToken(mockTokenId1) == uint256(oracle1InitialValue)
+        // Make sure the values are updated, start clean
+        localRelayer.execute();
+
+        // The initial value is the value we just defined
+        assertEq(
+            collybus.valueForToken(localTokenId),
+            uint256(initialValue),
+            "We should have the initial value"
+        );
+
+        // Update the oracle with a new value under the threshold limit
+        int256 secondValue = initialValue +
+            (initialValue * int256(thresholdPercentage)) /
+            100_00 -
+            1;
+        localOracle.givenQueryReturnResponse(
+            abi.encodePacked(IOracle.value.selector),
+            MockProvider.ReturnData({
+                success: true,
+                data: abi.encode(secondValue, true)
+            }),
+            false
+        );
+
+        // Relayer should not need to update values since it's below the thresold
+        assertTrue(localRelayer.check() == false, "Should not update values");
+
+        // Execute the relayer
+        localRelayer.execute();
+
+        // Make sure the new value was not pushed into Collybus
+        assertEq(
+            collybus.valueForToken(localTokenId),
+            uint256(initialValue),
+            "Collybus should not have been updated"
+        );
+    }
+
+    function test_Execute_UpdatesRatesInCollybusWhenDeltaIsAboveThreshold()
+        public
+    {
+        // Threshold percentage
+        uint256 thresholdPercentage = 50_00; // 50%
+
+        // TokenId
+        bytes32 localTokenId = bytes32("percentage_threshold_token_test");
+
+        // Create a local relayer
+        Relayer localRelayer = new Relayer(address(collybus), relayerType);
+
+        // Create a local oracle
+        MockProvider localOracle = new MockProvider();
+
+        // Set the value returned by the Oracle.
+        int256 initialValue = 100;
+        localOracle.givenQueryReturnResponse(
+            abi.encodePacked(IOracle.value.selector),
+            MockProvider.ReturnData({
+                success: true,
+                data: abi.encode(initialValue, true)
+            }),
+            false
+        );
+
+        // Add oracle with a thresold percentage
+        localRelayer.oracleAdd(
+            address(localOracle),
+            localTokenId,
+            thresholdPercentage
+        );
+
+        // Make sure the values are updated, start from `initialValue`
+        localRelayer.execute();
+
+        // The initial value is the value we just defined
+        assertEq(
+            collybus.valueForToken(localTokenId),
+            uint256(initialValue),
+            "We should have the initial value"
+        );
+
+        // Update the oracle with a new value above the threshold limit
+        int256 secondValue = initialValue +
+            (initialValue * int256(thresholdPercentage)) /
+            100_00 +
+            1;
+        localOracle.givenQueryReturnResponse(
+            abi.encodePacked(IOracle.value.selector),
+            MockProvider.ReturnData({
+                success: true,
+                data: abi.encode(secondValue, true)
+            }),
+            false
+        );
+
+        // Relayer should need to update values since it's above the thresold
+        assertTrue(localRelayer.check() == true, "Should update values");
+
+        // Execute the relayer
+        localRelayer.execute();
+
+        // Make sure the new value was pushed into Collybus
+        assertEq(
+            collybus.valueForToken(localTokenId),
+            uint256(secondValue),
+            "Collybus should have the new value"
         );
     }